Rearrangement of package structure
I’ve create and move several package. The idea is to base the packages
based on the modules definition that we have. Each of the package have
the same sub-structure: entities, services, dtos and constants.

There are two packages that I couldn’t refactor. Even-though I’ve decided
to submit the changes in order to let everyone move from the new structure.