Change attributes of SnapshotEconomic: Integer to Decimal
The changes were made in the database and in the entity.
In addition, the data type in the schema was changed from
integer to number.

The property validator from number was modified to accept
values ​​without a comma.

The static property mappers were modified so that it could
convert number to Double.